requirements.txt
pytest==5.4.1
requests==2.23.0
selenium==3.141.0
pytest-xdist==1.32.0
allure-pytest==2.8.16